+++ core/src/ecs/components/Position.ts
@@ -159,6 +159,7 @@ export const PositionSystem: SystemBuilder<"PositionSystem"> = {
   init: (world) => ({
     id: "PositionSystem",
     query: ["position"],
+    priority: 8,
     onTick: (entities: Entity<Position>[]) => {
       entities.forEach(entity => {
