+++ core/src/ecs/systems/core/ControlSystem.ts
@@ -5,6 +5,7 @@ export const ControlSystem: SystemBuilder = ({
   init: (world) => ({
     id: "ControlSystem",
     query: ["controlling"],
+    priority: 2,
     onTick: (entities: Entity<Controlling>[]) => {
       entities.forEach((entity) => {
         const character = entity.components.controlling.getCharacter(world)
