+++ core/src/runtime/World.ts
@@ -173,7 +173,9 @@ export const World = ({ commands, games, systems, renderer, mode }: WorldProps):
       }
 
       // run system onTick
-      values(world.systems).forEach((system) => {
+
+      values(world.systems).sort((a, b)=> a.priority - b.priority).forEach((system) => {
+        // console.log(`running system ${system.id} prio ${system.priority}`)
         if (!isRollback || (isRollback && !system.skipOnRollback)) {
           if (!world.systems[system.id]) return
           system.onTick?.(filterEntities(system.query, values(world.entities)), isRollback)
